Add authorize for ProjectController

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Requests\ProjectRequest;
 use App\Models\Project;
+use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
@@ -21,24 +22,33 @@ class ProjectController extends Controller
     }
 
     public function create() {
-        return view('admin.projects.create');
+        $users = User::all();
+        return view('admin.projects.create', compact('users'));
     }
 
-    public function store(ProjectRequest $projectRequest, Project $project){
-        Auth::user()->projects()->create($projectRequest->all());
+    public function store(ProjectRequest $projectRequest){
+        $project = Auth::user()->createProjects()->create($projectRequest->all());
+        $project->users()->attach($projectRequest->users);
         return redirect()->route('admin.projects.index');
     }
 
     public function edit(Project $project) {
-        return view('admin.projects.edit', compact('project'));
+        $this->authorize('update', $project);
+        $project->load('users');
+        $users = User::all();
+        return view('admin.projects.edit', compact('project', 'users'));
     }
 
     public function update(ProjectRequest $projectRequest, Project $project) {
+        $this->authorize('update', $project);
         $project->update($projectRequest->all());
+        $project->users()->detach();
+        $project->users()->attach($projectRequest->users);
         return redirect()->route('admin.projects.index');
     }
 
     public function destroy(Project $project) {
+        $this->authorize('delete', $project);
         $project->delete();
         return back();
     }
